restructure scatterPlotComponent and using atan2 for calculationg PA
