+++ src/components/Shared/ScatterPlot/ScatterPlotComponent.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import {observer} from "mobx-react";
 import ReactResizeDetector from "react-resize-detector";
 import {Layer, Stage} from "react-konva";
-import {Colors} from "@blueprintjs/core";
 import {ChartArea} from "chart.js";
 import {PlotContainerComponent} from "components/Shared/LinePlot/PlotContainer/PlotContainerComponent";
 import {ToolbarComponent} from "components/Shared/LinePlot/Toolbar/ToolbarComponent";
@@ -11,11 +10,6 @@ import "./ScatterPlotComponent.css";
 
 @observer
 export class ScatterPlotComponent extends LinePlotComponent {
-    private pointDefaultColor = Colors.GRAY2;
-    private opacityInit = 1;
-    private opacityOutRange = 0.1;
-    private pointRadiuInit = 3;
-    private pointRadiu = 9;
 
     private getChartAreaWH(chartArea: ChartArea): { width: number, height: number } {
         if (chartArea && chartArea.right && chartArea.bottom) {
@@ -23,86 +17,6 @@ export class ScatterPlotComponent extends LinePlotComponent {
         } else {
             return {width: 0, height: 0};
         }
-    }
-
-    private getScatterColor(currentIndex: number, range: number, toColor: number, frequencyIncreases: boolean): string {
-        let percentage = currentIndex / range;
-        if (!frequencyIncreases) {
-            percentage = 1 - percentage;
-        }
-        let hue = (percentage * toColor);
-        return `hsla(${hue}, 100%, 50%, ${this.opacityInit})`;
-    }
-
-    private frequencyIncreases (data: {x: number, y: number, z?: number}[]): boolean {
-        const zFirst = data[0].z;
-        const zLast = data[data.length - 1].z;
-        if (zFirst > zLast) {
-            return false;
-        }
-        return true;
-    }
-
-    private fillColor(): Array<string> {
-        let scatterColors = [];
-        if (this.props.data && this.props.data.length && this.props.zIndex && this.props.interactionBorder) {
-            let xlinePlotRange = this.props.interactionBorder;
-            const outOfRangeColor = `hsla(0, 0%, 50%, ${this.opacityOutRange})`;
-            const zOrder = this.frequencyIncreases(this.props.data);
-            const dataLength = this.props.data.length;
-            const colorRangeEnd = this.props.colorRangeEnd;
-            this.props.data.forEach((data, i) => {
-                let pointColor = this.pointDefaultColor;
-                let outRange = true;
-                if (data.z >= xlinePlotRange.xMin && data.z <= xlinePlotRange.xMax) {
-                    outRange = false;
-                }
-                pointColor = outRange ? outOfRangeColor : this.getScatterColor(i, dataLength, colorRangeEnd, zOrder);
-                scatterColors.push(pointColor);
-            });
-        }
-        return scatterColors;
-    }
-
-    private closestChannel(channel: number, data: Array<{x: number, y: number, z?: number}>): number {
-        var mid;
-        var lo = 0;
-        var hi = data.length - 1;
-        while (hi - lo > 1) {
-            mid = Math.floor ((lo + hi) / 2);
-            if (data[mid].z < channel) {
-                lo = mid;
-            } else {
-                hi = mid;
-            }
-        }
-        if (channel - data[lo].z <= data[hi].z - channel) {
-            return data[lo].z;
-        }
-        return data[hi].z;
-    }
-
-    private setPointRadius(): Array<number> {
-        let pointRadius = [];
-        if (this.props.data && this.props.data.length && this.props.zIndex && this.props.channel) {
-            let channelCurrent = this.props.channel.channelCurrent;
-            let channelHovered = this.props.channel.channelHovered;
-            pointRadius = Array(this.props.data.length).fill(this.pointRadiuInit);
-            if (channelCurrent) {
-                let close = channelCurrent;
-                if (channelHovered) {
-                    close = this.closestChannel(channelHovered, this.props.data);
-                }
-                const scatterData = this.props.data;
-                for (let index = 0; index < scatterData.length; index++) {
-                    const points = scatterData[index];
-                    if (points.z === channelCurrent || points.z === close) {
-                        pointRadius[index] = this.pointRadiu;
-                    }
-                }
-            }
-        }
-        return pointRadius;
     }
 
     private resizeData(): { xMin: number, xMax: number, yMin: number, yMax: number } {
@@ -146,8 +60,6 @@ export class ScatterPlotComponent extends LinePlotComponent {
                     chartAreaUpdated={this.updateChart}
                     width={this.width}
                     height={this.height}
-                    dataBackgroundColor={this.props.colorRangeEnd ? this.fillColor() : []}
-                    pointRadiusSet={this.setPointRadius()}
                     xMin={axisRange.xMin}
                     xMax={axisRange.xMax}
                     yMin={axisRange.yMin}
@@ -164,6 +76,7 @@ export class ScatterPlotComponent extends LinePlotComponent {
                     onWheel={this.onStageWheel}
                 >
                     <Layer>
+                        {this.genLines()}
                         {this.genBorderRect()}
                     </Layer>
                 </Stage>
